Placeholder Backend + Modified Structure
I'm uploading a placeholder backend folder with files from a previous project. We can modify these later or remove them but it's better to have the structure and example there. Front end is now in a separate folder. I also had to change the gitignore structure a bit to make it work